Fix colour command erroneously defined on client side.
